requirements.txt
opencv-python
numpy
pathlib
tqdm
python-magic